Add files via upload
Breaking News Radar is an innovative website that selects ,organize information about breaking news events from across the internet. When you search for an incident, it presents a comprehensive overview with summarized news, visuals, a map of the location, and a historical timeline. This allows you to get a complete, multi-faceted understanding of any major event quickly and efficiently in one place.